Check for delayed_job in a run before using it!
This was occasionally breaking cancel when a run had a delayed job id, but the
job itself had been cleaned up.

+++ app/models/taverna_player/run.rb
@@ -49,10 +49,12 @@ module TavernaPlayer
     def cancel
       return if finished? || cancelled?
 
-      Delayed::Job.transaction do
-        if delayed_job.locked_by.nil?
-          delayed_job.destroy
-          update_attribute(:status_message, "Cancelled")
+      unless delayed_job.nil?
+        Delayed::Job.transaction do
+          if delayed_job.locked_by.nil?
+            delayed_job.destroy
+            update_attribute(:status_message, "Cancelled")
+          end
         end
       end
